embedding/requirements.txt
fastapi
uvicorn[standard]
python-multipart
numpy
essentia-tensorflow